Cluster Description

You must specfiy the type of matrix you are inputting to cluster your object and we support three types: the path to your phylip and column distance file, or a sparse matrix.

Usage

cluster(
  cutoff,
  method,
  count_table,
  simularity_matrix = FALSE,
  random_seed = 123,
  ...
)

Arguments

cutoff

A cutoff value.

method

The type of cluster you wish to conduct; furthest, nearest, average, weighted.

count_table

A table of names and the given abundance per group.

simularity_matrix

are you using a simularity matrix or distance matrix

random_seed

you can set your own random seed for consistent results, if not it will be set to 123

...

Either your phylip file or column file path, or a sparse distance matrix

Value

A string of the given cluster.

Details

Detailed description of the function.

Examples

 # Using a sparse matrix
 i_values <- as.integer(1:100)
 j_values <- as.integer(sample(1:100, 100, TRUE))
 x_values <- as.numeric(runif(100, 0, 1))
 s_matrix <- Matrix::spMatrix(nrow=max(i_values), 
                              ncol=max(i_values), 
                              i=i_values, 
                              j=j_values, 
                              x=x_values)

 # Creating a count table using the sparse matrix
 count_table_sparse <- data.frame(sequence=as.character(i_values), 
                                 total=rep(1,times=100))
 # furthest method
 cluster_results <- cluster(cutoff=0.2, count_table = count_table_sparse, 
                            sparse_matrix=s_matrix, method="furthest")

 # With a phylip file and nearest methods
 count_table <- read.delim(example_path("amazon1.count_table"))
 amazon_data_phylip <- cluster(phylip_path=example_path("98_sq_phylip_amazon.dist"),
                               count_table = count_table, method="nearest", cutoff = 0.2)

 # With a column file and average methods 
 amazon_data_column <- cluster(column_path=example_path("96_sq_column_amazon.dist"),
                               count_table = count_table, method="average", cutoff = 0.2)

 # Weighted method
 amazon_data_column <- cluster(column_path=example_path("96_sq_column_amazon.dist"),
                               count_table = count_table, method="weighted", cutoff = 0.2)