Read Dist Description

Detailed description of the function.

Usage

read_dist(distance_file, count_table, cutoff, is_simularity_matrix)

Arguments

distance_file

Either a phylip or column distance file, or a sparse matrix.

count_table

A table of names and the given abundance per group.

cutoff

The value you wish to use as a cutoff when clustering. opti, furthest, nearest, average, or weighted.

is_simularity_matrix

are you using a simularity matrix or distance matrix?

Value

A distance object that contains all your distance information

Examples

 # Convert Phylip or column file to a sparse matrix
 library(Matrix) # spMatrix requires the use of the matrix library
 i_values <- as.integer(1:100)
 j_values <- as.integer(sample(1:100, 100, TRUE))
 x_values <- as.numeric(runif(100, 0, 1))
 s_matrix <- spMatrix(nrow=max(i_values),
                              ncol=max(i_values),
                              i=i_values,
                              j=j_values,
                              x=x_values)

 column_path <- example_path("amazon_column.dist")
 phylip_path <- example_path("amazon_phylip.dist")
 count_table <- read_count(example_path("amazon.count_table"))

data_column <- read_dist(column_path, count_table, 0.2, FALSE)
#> This is a column file. Processing now...
data_phylip <- read_dist(phylip_path, count_table, 0.2, FALSE)
#> This is a phylip file. Processing now...
data_sparse <- read_dist(s_matrix, count_table, 0.2, FALSE)